Time: 5 ms (83.95%), Space: 22.9 MB (73.87%) - LeetHub

+++ 0907-koko-eating-bananas/0907-koko-eating-bananas.cpp
@@ -7,7 +7,7 @@ public:
         while(s<=e){
             long long mid = s+(e-s)/2;
             long long count = 0;
-            for(long long i:piles){
+            for(int i:piles){
                 if(i%mid == 0){
                     count += i/mid;
                 }
@@ -23,6 +23,6 @@ public:
                 s = mid+1;
             }
         }
-        return ans;
+        return (int)ans;
     }
 };
